+++ Soggins/BotchSoundInfo.cs
@@ -1,0 +1,8 @@
+﻿using Nickel;
+
+namespace Shockah.Soggins;
+
+internal record struct BotchSoundInfo(
+	ISoundEntry Entry,
+	double ActionDelay = 0.2
+);