+++ Soggins/docs/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 
-## Upcoming release
+## 1.9.0
+Released 23 February 2026.
 
 * Changed the Stop It! card, effectively merging it with Zen.
 * Replaced the Zen card with the Mood Swing card.